+++ wxc_sdk/telephony/location/__init__.py
@@ -232,7 +232,8 @@ class TelephonyLocationApi(ApiChild, base='telephony/config/locations'):
         :type org_id: str
         :return:
         """
-        data = settings.json(exclude={'location_id', 'user_limit', 'default_domain'})
+        data = settings.json(exclude={'location_id', 'name', 'user_limit', 'default_domain', 'subscription_status',
+                                      'e911_setup_required'})
         params = org_id and {'orgId': org_id} or None
         url = self.ep(location_id)
         self.put(url=url, data=data, params=params)